requirements.txt
educhain>=0.3.10
mcp>=0.9.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0